backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
reportlab==4.1.0
qrcode[pil]==7.4.2
pillow==10.2.0
httpx==0.27.0
motor==3.3.2
beanie==1.25.0
pydantic[email]==2.6.1
boto3>=1.34.0
